Update README.md
The Timekeeper provides the first module for time recording and management of employees and workers in a company. Other building blocks will follow. React is used as the front end and the back end consists of python with SQLite for persistent data storage. Personal, sensitive data is stored in encrypted form.